Refactor code and add item-counter test
