+++ .hintrc
