+++ src/__test__/itemcounter.test.js
@@ -1,0 +1,9 @@
+/**
+ * @jest-environment jsdom
+ */
+import itemCounter from '../itemCount.js';
+
+test('Items counts is', async () => {
+  const total = await itemCounter();
+  expect(total).toBe(24);
+});
