+++ src/style.css
@@ -8,7 +8,7 @@ body {
 }
 
 header {
-  border: 1px solid black;
+  border-bottom: 1px solid black;
   height: 90px;
   display: flex;
   align-items: center;
